Issues fixed

+++ src/main/java/com/smartstay/smartstay/Wrappers/Electricity/ElectricityUsageMapper.java
@@ -28,6 +28,6 @@ public class ElectricityUsageMapper implements Function<RoomInfoForEB, Electrici
                 0.0,
                 0.0,
                 0.0,
-                roomInfoForEB.noOfTenants());
+                roomInfoForEB.noOfTenants().intValue());
     }
 }
